+++ src/rumbleship-beeline.ts
@@ -4,6 +4,7 @@ import {
   HoneycombSchema,
   IAsyncTracker
 } from './honeycomb.interfaces';
+import { SamplerPipeline } from './sampler-pipeline';
 export class RumbleshipBeeline {
   private static beeline: any; // The wrapped beeline from `require('honeycomb-beeline')`;
   static TrackedContextbyContextId: Map<string, any> = new Map();
@@ -20,7 +21,8 @@ export class RumbleshipBeeline {
     if (this.initialized) {
       throw new Error('RumbleshipBeeline already initialized as a singleton. Cannot reinitialize');
     }
-    this.beeline = configureBeeline(config);
+    const sampler = new SamplerPipeline();
+    this.beeline = configureBeeline({ samplerHook: sampler.sample, ...config });
     this.initialized = true;
   }
 
@@ -81,7 +83,7 @@ export class RumbleshipBeeline {
     } catch (error) {
       if (error.extensions) {
         for (const [k, v] of Object.entries(error.extensions)) {
-          this.addContext({ [`app.gql.error.extensions.${k}`]: v });
+          this.addTraceContext({ [`gql.error.extensions.${k}`]: v });
         }
       }
       // Is this right?
@@ -109,7 +111,7 @@ export class RumbleshipBeeline {
           innerValue = fn(span);
         } catch (error) {
           // catch errors here and update the span
-          this.addContext({
+          this.addTraceContext({
             error: `${error}`,
             'error.message': error.message,
             'error.stack': error.stack
@@ -117,7 +119,7 @@ export class RumbleshipBeeline {
 
           if (error.extensions) {
             for (const [k, v] of Object.entries(error.extensions)) {
-              this.addContext({ [`app.gql.error.extensions.${k}`]: v });
+              this.addTraceContext({ [`gql.error.extensions.${k}`]: v });
             }
           }
 
@@ -137,14 +139,16 @@ export class RumbleshipBeeline {
           (innerValue as Promise<T>)
             .catch((error: Error) => {
               // catch errors here and update the span
-              this.addContext({
+              this.addTraceContext({
                 error: `${error}`,
                 'error.message': error.message,
                 'error.stack': error.stack
               });
               if ((error as any).extensions) {
                 for (const [k, v] of Object.entries((error as any).extensions)) {
-                  this.addContext({ [`app.gql.error.extensions.${k}`]: v });
+                  this.addTraceContext({
+                    [`gql.error.extensions.${k}`]: v
+                  });
                 }
               }
               throw error;
@@ -197,6 +201,7 @@ export class RumbleshipBeeline {
       parentSpanId,
       dataset
     );
+    this.addContext({ gae_version: process.env.GAE_VERSION });
     RumbleshipBeeline.TrackedContextbyContextId.set(
       this.context_id,
       RumbleshipBeeline.HnyTracker?.getTracked()
@@ -244,8 +249,21 @@ export class RumbleshipBeeline {
   runWithoutTrace<T>(fn: () => T): T {
     return RumbleshipBeeline.beeline.runWithoutTrace(fn);
   }
+  /**
+   *
+   * @param context Add keys+values of an object to JUST the current span
+   * @note you probably want `addTraceContext()` to propagate your metadata to all children.
+   */
   addContext(context: object): void {
     return RumbleshipBeeline.beeline.addContext(context);
+  }
+  /**
+   *
+   * @param context Add keys+values of object to the current span AND ALL CHILD SPANS
+   *  Keys are automatically prefixed with `app.`
+   */
+  addTraceContext(context: object): void {
+    return RumbleshipBeeline.beeline.addTraceContext(context);
   }
   removeContext(context: object): void {
     return RumbleshipBeeline.beeline.removeContext(context);